Simplify number handling by removing IntegerLiteral type
- Remove IntegerLiteral type, using only NumberLiteral (float64) for all numeric values
- Simplify parseNumber to directly use ParseFloat without redundant integer parsing
- Add type conversion logic for float64/int64 interoperability
- Update all tests to use NumberLiteral instead of IntegerLiteral

+++ query/sql_test.go
@@ -19,7 +19,7 @@ func TestToSql(t *testing.T) { //nolint:funlen
 			// Simple equality using colon (converted to "=")
 			input:    "status:200",
 			want:     "SELECT * FROM dummy_table WHERE status = ?",
-			wantArgs: []any{int64(200)},
+			wantArgs: []any{float64(200)},
 		},
 		{
 			// Floating-point comparison with ">"
@@ -31,25 +31,25 @@ func TestToSql(t *testing.T) { //nolint:funlen
 			// Boolean AND between two conditions.
 			input:    "status:200 and eps < 0.003",
 			want:     "SELECT * FROM dummy_table WHERE (status = ? AND eps < ?)",
-			wantArgs: []any{int64(200), 0.003},
+			wantArgs: []any{float64(200), 0.003},
 		},
 		{
 			// Boolean OR between two conditions.
 			input:    "status:200 or eps < 0.003",
 			want:     "SELECT * FROM dummy_table WHERE (status = ? OR eps < ?)",
-			wantArgs: []any{int64(200), 0.003},
+			wantArgs: []any{float64(200), 0.003},
 		},
 		{
 			// NOT operator applied to a field expression.
 			input:    "not status:200",
 			want:     "SELECT * FROM dummy_table WHERE NOT status = ?",
-			wantArgs: []any{int64(200)},
+			wantArgs: []any{float64(200)},
 		},
 		{
 			// Parenthesized expression.
 			input:    "(status:200 and eps<0.003)",
 			want:     "SELECT * FROM dummy_table WHERE (status = ? AND eps < ?)",
-			wantArgs: []any{int64(200), 0.003},
+			wantArgs: []any{float64(200), 0.003},
 		},
 		{
 			// Array literal conversion (using IN).
@@ -61,25 +61,25 @@ func TestToSql(t *testing.T) { //nolint:funlen
 			// Complex expression combining AND and a parenthesized array literal.
 			input:    "status:200 and eps<0.003 and (req.fields.ext:[\"jpg\", \"png\"])",
 			want:     "SELECT * FROM dummy_table WHERE ((status = ? AND eps < ?) AND req.fields.ext IN (?,?))",
-			wantArgs: []any{int64(200), 0.003, "jpg", "png"},
+			wantArgs: []any{float64(200), 0.003, "jpg", "png"},
 		},
 		{
 			// Greater than or equal operator.
 			input:    "cmp>=100",
 			want:     "SELECT * FROM dummy_table WHERE cmp >= ?",
-			wantArgs: []any{int64(100)},
+			wantArgs: []any{float64(100)},
 		},
 		{
 			// Less than or equal operator.
 			input:    "price<=50",
 			want:     "SELECT * FROM dummy_table WHERE price <= ?",
-			wantArgs: []any{int64(50)},
+			wantArgs: []any{float64(50)},
 		},
 		{
 			// Nested NOT with a parenthesized expression.
 			input:    "not (status:200 and eps < 0.003)",
 			want:     "SELECT * FROM dummy_table WHERE NOT (status = ? AND eps < ?)",
-			wantArgs: []any{int64(200), 0.003},
+			wantArgs: []any{float64(200), 0.003},
 		},
 		{
 			input: `name~"John"`,
